docker/requirements.txt
fastapi
boto3
starlette
llama-cpp-python
pydantic
uvicorn
requests